% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ayudantia.epg3308.R
\docType{package}
\name{ayudantia.epg3308}
\alias{ayudantia.epg3308}
\alias{ayudantia.epg3308-package}
\title{ayudantia.epg3308: Un package desarrollado para las Ayudantías}
\description{
Este package será incluido en algunas de las actividades realizadas
en las Ayudantías de EPG3308 - Computación Estadística
}
\author{
Esteban Rucán
}
