% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ejemplo_de_documentacion.R
\encoding{UTF-8}
\name{ejemplo_de_documentacion}
\alias{ejemplo_de_documentacion}
\title{Ejemplo de Documentación}
\usage{
ejemplo_de_documentacion()
}
\value{
Mensaje con un consejo
}
\description{
Ejemplo para los estudiantes
}
\details{
Esta función está hecha para que los estudiantes vean un ejemplo de documentación
}
\examples{
?ejemplo_de_documentacion

}
\author{
Esteban Rucán
}
