% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generar_estructura.R
\encoding{UTF-8}
\name{generar_estructura}
\alias{generar_estructura}
\title{Estructura básica de organización}
\usage{
generar_estructura()
}
\value{
Carpetas en el directorio de trabajo.
}
\description{
Puedes usar esta función para ordenar tu trabajo a través de crear carpetas
automáticamente en el directorio %$%  actual.
}
\details{
La buena estructuración de un espacio de trabajo es muy útil para mantener el orden y
la legitividad de tu trabajo.
}
\examples{
getwd()                # Obtener directorio.
dir()                  # Revisa las carpetas que hay actualmente en el directorio.
generar_estructura()   # Crear las carpetas.
dir()                  # Revisa las carpetas creadas.

}
\author{
Esteban Rucán.
}
