% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lista_objetos.R
\docType{data}
\name{lista_objetos}
\alias{lista_objetos}
\title{Lista con objetos de distinto tipo.}
\format{
Una lista con 4 objetos:
\describe{
  \item{data_frame}{Tibble 'diamonds' del package \code{ggplot2}}
  \item{matriz}{Una matriz de 100x100 con valores simulados desde una ditribución uniforme con parámetros -10 y 10}
  \item{sub_lista}{Una lista que contiene 20 vectores numéricos con 20 elementos cada uno, obtenidos desde una distribución exponencial con parámetro 20}
  \item{vector_palabras}{Un vector del tipo \code{character} con 100 elementos extraidos desde el package \code{stringr}}
}
}
\source{
\code{ggplot2}, \code{stringr}
}
\usage{
lista_objetos
}
\description{
Una lista creada para trabajar en la segunda ayudantía
}
\keyword{datasets}
