% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numero_de_ayudantia.R
\encoding{UTF-8}
\name{numero_de_ayudantia}
\alias{numero_de_ayudantia}
\title{Crear Plantilla de R Markdown}
\usage{
numero_de_ayudantia()
}
\value{
Archivo con extensión \code{.Rmd}.
}
\description{
Función que indica el número de ayudantía
}
\details{
Esta función está pensada para la plantilla de R Markdown de las ayudantías
}
\examples{
numero_de_ayudantia()
}
\author{
Esteban Rucán.
}
