% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_mean.R
\encoding{UTF-8}
\name{power_mean}
\alias{power_mean}
\title{Power Mean}
\usage{
power_mean(x, m, na.rm)
}
\arguments{
\item{x}{Un vector con valores numéricos.}

\item{m}{Potencia de la media generalizada. Si es 0, hace referencia a la media geométrica.}

\item{na.rm}{Valor Lógico, si es \code{TRUE}, excluye \code{NA} de los cálulos.}
}
\value{
Media generalizada según la potencia \code{m}.
}
\description{
Calcular la Media Generalizada de un vector y potencia \code{m}.
}
\details{
Las medias generalizadas, también conocidas como medias de Hölder, son una abstracción de las medias cuadráticas, aritméticas, geométricas y armónicas. Se definen y agrupan a través de la siguiente expresión:
(1 / n * sum(x ^ m)) ^ (1 / m).
}
\examples{
power_mean(x = c(1:100, NA), m = 2, na.rm = TRUE)
power_mean(3:10, 2, FALSE)
power_mean(50:60, 0, FALSE)

}
\author{
Esteban Rucán.
}
