% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usar_plantilla.R
\encoding{UTF-8}
\name{usar_plantilla}
\alias{usar_plantilla}
\title{Crear Plantilla de R Markdown}
\usage{
usar_plantilla()
}
\value{
Archivo con extensión \code{.Rmd}.
}
\description{
Puedes usar esta función para crear una plantilla de RMarkdown en el directorio de trabajo actual.
}
\details{
La buena estructuración de un documento mejora la comprensión del mismo.
}
\examples{
usar_plantilla()

}
\author{
Esteban Rucán.
}
