% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mensaje_de_bienvenida.R
\encoding{UTF-8}
\name{mensaje_de_bienvenida}
\alias{mensaje_de_bienvenida}
\title{Bienvenida al Curso EPG3308}
\usage{
mensaje_de_bienvenida()
}
\value{
Mensaje de bienvenida
}
\description{
Muestra un mensaje de bienvenida
}
\details{
Esta función ha sido diseñada para darles la bienvenida a los estudiantes del Curso EPG3308 y que el primer package que carguen les muestre un lindo mensaje.
}
\author{
Esteban Rucán
}
